update llm config
